vibe-studio/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-dotenv>=1.0
replicate>=0.26
requests>=2.31
PyYAML>=6.0.1
python-multipart>=0.0.9